+++ changelog.md
@@ -72,3 +72,6 @@ Remove not needed second expansion step in mergeOverlapping and remove the assoc
 
 Improve reporting of evaluation results and fix a bug that was causing the file list to be in the wrong order.
 
+* v11.1
+New output structure for functions. 
+Rename boundingBoxesFromLastSection to autoROI